Autoloader to PSR4. API classes one level up in dir structure. To be used in dev package in packagist
