+++ composer.json
@@ -13,8 +13,8 @@
         "php": ">=5.3.0"
     },
     "autoload": {
-        "psr-0": {
-            "Patreon": "src/"
+        "psr-4": {
+            "DevPatreon": "src/"
         }
     }
 }
